Reorganize CUDA's memory stuff into its own directory
