+++ agency/cuda/detail/memory/unique_ptr.hpp
@@ -2,7 +2,7 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/detail/unique_ptr.hpp>
-#include <agency/cuda/detail/allocator.hpp>
+#include <agency/cuda/detail/memory/allocator.hpp>
 
 namespace agency
 {
